Fix pretrained_distilbert.py

+++ Knowledge_Tracing/code/models/nlp_models/pretrained_distilbert.py
@@ -284,7 +284,7 @@ class PretrainedDistilBERT():
             encoding = output.to_tuple()[0]
             for problem_id, enc, attention in list(zip(self.texts_df['problem_id'].values[start:end],
                                                        encoding, attention_mask)):
-                self.encodings[problem_id] = F.normalize(mean_pool(enc, attention), p=2, dim=0).detach().numpy().float()
+                self.encodings[problem_id] = F.normalize(mean_pool(enc, attention), p=2, dim=0).detach().numpy()
                 print(self.encodings[problem_id])
             start = start + batch_size
         print(len(list(self.encodings.keys())))
